% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportInit.R
\name{appendBelowYAML}
\alias{appendBelowYAML}
\title{Add lines below the YAML header}
\usage{
appendBelowYAML(filepath, what)
}
\arguments{
\item{filepath}{file path}

\item{what}{character vector ines to add}
}
\value{
logical; \code{TRUE} if successful
}
\description{
Add lines below the YAML header
}
