% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{cLHS_subset}
\alias{cLHS_subset}
\title{Use cLHS to subset a data.frame using selected variables}
\usage{
cLHS_subset(i, n, non.id.vars)
}
\arguments{
\item{i}{a \code{data.frame}}

\item{n}{number of cLHS samples (rows) to draw}

\item{non.id.vars}{variables that are non-ID columns}
}
\value{
a subset \code{data.frame} corresponding to selected cLHS samples (rows)
}
\description{
Use cLHS to subset a data.frame using selected variables
}
