% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportInit.R
\name{copyPath}
\alias{copyPath}
\title{Copy a file from source to output directory}
\usage{
copyPath(fname, srcDir, outputDir, overwrite = FALSE)
}
\arguments{
\item{fname}{file name}

\item{srcDir}{source directory}

\item{outputDir}{output directory}

\item{overwrite}{overwrite? default: \code{FALSE}}
}
\value{
logical; result of \code{file.copy}
}
\description{
Copy a file from source to output directory
}
