% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{custom.bwplot}
\alias{custom.bwplot}
\title{Title}
\usage{
custom.bwplot(x, coef = NA, do.out = FALSE)
}
\arguments{
\item{x}{vector of values to summarize}

\item{coef}{Moran's I associated with the current raster}

\item{do.out}{not used}
}
\value{
a list containing elements: \code{stats}, \code{n}, \code{conf}, and \code{out}
}
\description{
Title
}
