% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportInit.R
\name{defineInCodeChunk}
\alias{defineInCodeChunk}
\title{Define a parameter in a code chunk}
\usage{
defineInCodeChunk(filepath, param.name, param.value)
}
\arguments{
\item{filepath}{File to add code chunk to}

\item{param.name}{Parameter name}

\item{param.value}{Parameter value}
}
\value{
logical; \code{TRUE} if successful
}
\description{
Define a parameter in a code chunk
}
