% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportInit.R
\name{defineInYAMLHeader}
\alias{defineInYAMLHeader}
\title{Define a parameter in the YAML header}
\usage{
defineInYAMLHeader(filepath, param.name, param.value)
}
\arguments{
\item{filepath}{File to add code chunk to}

\item{param.name}{Parameter name}

\item{param.value}{Parameter value}
}
\value{
logical; \code{TRUE} if successful
}
\description{
Define a parameter in the YAML header
}
