% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{findSafeVars}
\alias{findSafeVars}
\title{Return variables with sufficient variation for cLHS}
\usage{
findSafeVars(x, id, tol = 1e-05)
}
\arguments{
\item{x}{data.frame in wide format}

\item{id}{vector of IDs variables to exclude from SD test, first element is the group ID}

\item{tol}{tolerance for near-0 SD test; Default: \code{1e-5}}
}
\value{
A character vector of "safe" column names
}
\description{
Datasets that are nearly or completely invariant can pose issues for the cLHS algorithm. This function identifies vairables that have standard deviation smaller than a specified tolerance. Default: \code{1e-5}
}
