% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{kdeContours}
\alias{kdeContours}
\title{Calculate kernel density contour lines at specified probability levels
Calculate kernel density contour lines at specified probability levels with \code{MASS:kde2d} and display with \code{graphics::contour}}
\usage{
kdeContours(i, id, prob, cols, m, ...)
}
\arguments{
\item{i}{a data.frame containing unique ID, \code{x}, \code{y}}

\item{id}{a unique ID column name}

\item{prob}{a vector of probability levels}

\item{cols}{a vector of colors}

\item{m}{unique levels of the ID column (used to match colors)}

\item{...}{additional arguments to \code{graphics::contour}}
}
\value{
estimated kernel density contours
}
\description{
Calculate kernel density contour lines at specified probability levels
Calculate kernel density contour lines at specified probability levels with \code{MASS:kde2d} and display with \code{graphics::contour}
}
