% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listReports.R
\name{listReports}
\alias{listReports}
\title{List names of available reports}
\usage{
listReports(showFullPaths = FALSE)
}
\arguments{
\item{showFullPaths}{Show full paths to files? Default: \code{FALSE}}
}
\value{
a \code{data.frame} containing report names, versions and descriptions for reports available in the currently installed version of the soilReports package.
}
\description{
List names of available reports
}
\examples{

listReports()

listReports(showFullPaths = TRUE)

}
