% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pIndex}
\alias{pIndex}
\title{Indexing for Plotting and Printing}
\usage{
pIndex(x, interval = 4)
}
\arguments{
\item{x}{a \code{data.frame}, \code{SoilProfileCollection} or \code{table}}

\item{interval}{a value specifying the interval length desired; Default: \code{4}}
}
\value{
a numeric vector
}
\description{
This function creates an index to iterate over when plotting or printing large objects.
}
\examples{
x <- as.data.frame(matrix(1:100, ncol = 10))
pIndex(x, interval = 3)

}
