% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prettySummary}
\alias{prettySummary}
\title{Pretty Quantile Printing}
\usage{
prettySummary(x, p = c(0, 0.25, 0.5, 0.75, 1), n = TRUE, signif = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{p}{a numeric vector of percentiles}

\item{n}{a logical value indicating whether the vector of percentiles should be appended with the number of observations}

\item{signif}{a logical value indicating whether the percentiles should be rounded to the precision of the data}
}
\value{
a character value of quantiles and optionally the number of observations
}
\description{
This function generates a pretty quantile summary for printing.
}
\examples{
x <- 1.1:10.1
prettySummary(x)

}
\author{
Stephen Roecker
}
