% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportInit.R
\name{reportInit}
\alias{reportInit}
\alias{copyReport}
\alias{reportUpdate}
\title{Copy default configuration file and report contents to new directory}
\usage{
reportInit(
  reportName,
  outputDir = NULL,
  overwrite = FALSE,
  updateReport = FALSE
)
}
\arguments{
\item{reportName}{Name of report, as found in \code{listReports.} Format: \code{directory/reportName}.}

\item{outputDir}{Directory to create report instance}

\item{overwrite}{Overwrite existing directories and files? Default FALSE}

\item{updateReport}{Only update core report files, leaving configuration unchanged? Specific settings are report-dependent and set in the setup.R manifest.}
}
\value{
A time-stamped report instance created in \code{outputDir}, and a message summarizing the action(s) completed.
}
\description{
\code{reportInit} allows creation new report instances, or updates, from the soilReports R package. soilReports is a container for reports and convenience functions for soil data summary, comparison, and evaluation reports used mainly by USDA-NRCS staff.
}
