% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSetup.R
\name{reportSetup}
\alias{reportSetup}
\title{Install packages needed for a report}
\usage{
reportSetup(reportName, upgrade = FALSE)
}
\arguments{
\item{reportName}{Name of report, as found in \code{listReports.} Format: \code{directory/reportName}.}

\item{upgrade}{Upgrade CRAN packages? Default: \code{FALSE}}
}
\value{
Installed packages from CRAN and GitHub in user library, as specified in report-specific manifest.
}
\description{
Install packages needed for a report
}
