\name{soilReports-package}
\alias{soilReports-package}
\alias{soilReports}
\alias{installRprofile}
\alias{listReports}
\alias{reportInit}
\alias{reportSetup}
\docType{package}
\title{
Soil Reports
}
\description{R package container and convenience functions for soil data summary, comparison, and evaluation reports used mainly by NRCS staff.

\href{https://github.com/ncss-tech/soilReports}{Instructions and project homepage.}

\author{Dylan E. Beaudette <dylan.beaudette@ca.usda.gov>}
\keyword{ package }


