% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installRprofile.R
\name{installRprofile}
\alias{installRprofile}
\title{Install USDA-NRCS SPSD CCE user .Rprofile}
\usage{
installRprofile(overwrite = FALSE)
}
\arguments{
\item{overwrite}{overwrite? Default \code{FALSE}}
}
\value{
\code{source()} a new .Rprofile set up to redirect R library paths
}
\description{
Install USDA-NRCS SPSD CCE user .Rprofile
}
